requirements.txt
django>=5.2.1
djangorestframework>=3.16.0
django-cors-headers>=4.7.0
python-jose[cryptography]>=3.4.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.1.0
psycopg2-binary>=2.9.10
requests>=2.31.0
graphene-django>=3.0.0
whitenoise>=6.0.0
pyotp>=2.8.0
qrcode>=7.3.1
django-allauth>=0.58.0
PyJWT>=2.0.0 